✨feat : hook요소 파일 분리 및 적용
